requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.9.4
aniso8601==9.0.1
apispec==6.3.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.4
base58==2.1.1
bcrypt==4.0.1
bitarray==2.7.3
blinker==1.6.2
cachelib==0.9.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cryptography==36.0.1
cytoolz==0.12.1
Deprecated==1.2.14
distlib==0.3.6
dominate==2.7.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==18.11.1
filelock==3.12.2
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.0.2
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-HTTPAuth==4.7.0
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
flask-restx==1.1.0
Flask-Share==0.1.1
flask-smorest==0.42.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.0.1
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
limits==3.5.0
lru-dict==1.1.8
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mdurl==0.1.2
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
oauthlib==3.2.2
ordered-set==4.1.0
packaging==23.0
parsimonious==0.8.1
passlib==1.7.4
Pillow==9.5.0
platformdirs==3.6.0
pluggy==1.0.0
protobuf==3.20.3
psycopg2-binary==2.9.6
pycodestyle==2.11.0
pycparser==2.21
pycryptodome==3.17
Pygments==2.15.1
PyJWT==2.6.0
PyMySQL==1.0.2
pyperclip==1.8.2
pypng==0.20220715.0
pyrsistent==0.19.3
pytest==7.2.2
pytest-flask==1.2.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML @ file:///private/tmp/pyyaml-20230717-6270-1iiwi1j/PyYAML-6.0.1
qr==0.6.0
qrcode==7.4.2
redis==4.5.5
requests==2.27.1
requests-oauthlib==1.3.1
rich==13.4.2
rlp==2.0.1
shortuuid==1.0.11
six==1.16.0
SQLAlchemy==2.0.4
tk==0.1.0
tomli==2.0.1
toolz==0.12.0
typing_extensions==4.5.0
urllib3==1.26.15
varint==1.0.2
virtualenv==20.23.1
visitor==0.1.3
web3==5.27.0
webargs==8.3.0
websockets==9.1
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.8.2
